Add files via upload
The program will receive from the user various alphabetic characters. 
If the user wants to know the number of alphabetic characters he has entered. he will
have to input the following two characters: ** 
The alphabetic characters between the ** will be counted. The program will be terminated
when the character(*) will be inputed for a second time